Added tests and factory models for api

+++ news/views.py
@@ -128,7 +128,7 @@ class NewsList(viewsets.ModelViewSet):
             if queryset.count() != 0:
                 return Response(data=serializer.data, status=200)
             else:
-                return Response(data="object not found", status=status.HTTP_400_BAD_REQUEST)
+                return Response(data="object not found", status=status.HTTP_404_NOT_FOUND)
 
 
 class ApprovedNewsList(viewsets.ModelViewSet):
@@ -237,12 +237,12 @@ class ApprovedNewsList(viewsets.ModelViewSet):
         Возврощает новость по custom_url
         """
         if custom_url is not None:
-            queryset = News.objects.all().filter(custom_url=custom_url, is_approved=True)
+            queryset = News.objects.filter(custom_url=custom_url, is_approved=True)
             serializer = SingleNewsSerializer(queryset, many=True)
             if queryset.count() != 0:
                 return Response(data=serializer.data, status=200)
             else:
-                return Response(data="object not found", status=status.HTTP_400_BAD_REQUEST)
+                return Response(data="object not found", status=status.HTTP_404_NOT_FOUND)
 
     @swagger_auto_schema(
         responses={
@@ -468,7 +468,7 @@ class TagsList(viewsets.ModelViewSet):
             if queryset.count() != 0:
                 return Response(data=serializer.data, status=200)
             else:
-                return Response(data="object not found", status=status.HTTP_400_BAD_REQUEST)
+                return Response(data="object not found", status=status.HTTP_404_NOT_FOUND)
 
 
 class CategoriesList(viewsets.ModelViewSet):
@@ -513,7 +513,7 @@ class CategoriesList(viewsets.ModelViewSet):
             if queryset.count() != 0:
                 return Response(data=serializer.data, status=200)
             else:
-                return Response(data="object not found", status=status.HTTP_400_BAD_REQUEST)
+                return Response(data="object not found", status=status.HTTP_404_NOT_FOUND)
 
 
 class AuthorList(viewsets.ModelViewSet):
@@ -630,7 +630,7 @@ class NewsUserList(viewsets.ModelViewSet):
     )
     def retrieve(self, request, pk=None):
         if pk is not None:
-            queryset = NewsUser.objects.get(google_id=pk)
+            queryset = NewsUser.objects.filter(google_id=pk).first()
             serializer = NewsUserSerializer(queryset)
             return Response(data=serializer.data, status=200)
             # else:
@@ -743,10 +743,14 @@ class CommentList(viewsets.ModelViewSet):
         except:
             return Response(status=status.HTTP_404_NOT_FOUND, data="News not found!")
         try:
+            comment = Comment.objects.get(id=comment_id)
+        except:
+            return Response(status=status.HTTP_404_NOT_FOUND, data="Comment not found!")
+        try:
             user = NewsUser.objects.get(email=user_email)
         except:
             return Response(status=status.HTTP_404_NOT_FOUND, data="User not found!")
-        if news and user:
+        if news and user and comment:
             Comment.objects.filter(id=comment_id, news__id=news__id, author__email=user_email).delete()
 
         return Response(status=status.HTTP_200_OK, data="Succesfully deleted!")
